will REFACTOR: some components into more appropriate folders and will display the resident form on resident page
