+++ solar-panel-planner/src/pages/ResidentPage.jsx
@@ -1,16 +1,21 @@
-import { useNavigate } from "react-router-dom"
+import { useNavigate } from "react-router-dom";
+import ResidentForm from "../ui/ResidentForm";
 
 function ResidentPage() {
-const navigate = useNavigate()
+  const navigate = useNavigate();
 
-    return (
-        <div>
-            ResidentPage
-            <button onClick={()=> navigate("/")} className="bg-primaryGreen text-white text-xl font-semibold px-9 py-5 rounded-lg shadow-md hover:bg-secondaryGreen">
+  return (
+    <div>
+      ResidentPage
+      <button
+        onClick={() => navigate("/")}
+        className="bg-primaryGreen text-white text-xl font-semibold px-9 py-5 rounded-lg shadow-md hover:bg-secondaryGreen"
+      >
         Back
       </button>
-        </div>
-    )
+      <ResidentForm />
+    </div>
+  );
 }
 
-export default ResidentPage
+export default ResidentPage;
