+++ solar-panel-planner/src/utils/Modal.jsx
